+++ Sources/Core/ModernAVPlayer.swift
@@ -56,10 +56,21 @@ public final class ModernAVPlayer: NSObject, ModernAVPlayerExposable {
         return context.currentMedia
     }
 
+    /// Player's current time
+    public var currentTime: Double {
+        return context.currentTime
+    }
+
     /// Enable/Disable loop on the current media
     public var loopMode: Bool {
         get { return context.loopMode }
         set { context.loopMode = newValue }
+    }
+
+    /// Remote command preset
+    public var remoteCommands: [ModernAVPlayerRemoteCommand]? {
+        get { return context.remoteCommands }
+        set { context.remoteCommands = newValue }
     }
 
     // MARK: - Input
@@ -87,7 +98,7 @@ public final class ModernAVPlayer: NSObject, ModernAVPlayerExposable {
         
         defer {
             if config.useDefaultRemoteCommand {
-                ModernAVPlayerRemoteCommandCenter(player: self)
+                remoteCommands = ModernAVPlayerRemoteCommandFactory(player: self).defaultCommands
             }
         }
     }